restructuring and consistancy
