+++ src/components/CourseComp.tsx
@@ -1,16 +1,16 @@
 import React from "react";
-import { Card } from "react-bootstrap/";
+import { Card, Accordion } from "react-bootstrap/";
 import "../css/courses.css";
-import "../assets/courses.json";
+import "../assets/courses";
 import { Course } from "../interfaces/course";
 
 
 function CourseComp({ course1 }: {
-    course1: Course
+    course1: Course 
     }):  JSX.Element {
     
     return (
-        <Card className="card" >
+        <Card className="card">
             <Card.Body>
                 <Card.Title>{ course1.name }</Card.Title>
                 <Card.Text>
@@ -20,8 +20,18 @@ function CourseComp({ course1 }: {
                 </Card.Text>
             </Card.Body>
         </Card>
-    
     );
 }
 
 export default CourseComp;
+
+/*
+            <Accordion defaultActiveKey="0" flush>
+                <Accordion.Item eventKey="0">
+                    <Accordion.Header>Accordion Item #1</Accordion.Header>
+                    <Accordion.Body>
+                        bruh
+                    </Accordion.Body>
+                </Accordion.Item>
+            </Accordion>
+*/
