+++ src/index.tsx
@@ -4,11 +4,13 @@ import "./index.css";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 import "../node_modules/bootstrap/dist/css/bootstrap.min.css";
-
+import CourseComp from "./components/CourseComp";
+import courseData from "./assets/courses";
 
 ReactDOM.render(
     <React.StrictMode>
         <App />
+        
     </React.StrictMode>,
     document.getElementById("root")
 );
